refactor(client): split event_source.py into a package
The single 1422-line event_source.py grew to mix three concerns: the
abstract event-source contract, the inbound gossip wire-format parser,
and the live network orchestrator. This PR splits it into a focused
package without rewriting any documentation or changing any behavior.

New layout:

- event_source/protocol.py: EventSource Protocol, GossipMessageError,
  SUPPORTED_PROTOCOLS allow-list (the dependency-light shared surface).
- event_source/gossip.py: GossipHandler and read_gossip_message (the
  inbound gossipsub wire-format parser).
- event_source/live.py: LiveNetworkEventSource (the orchestrator that
  owns connections, streams, and the gossipsub behavior).
- event_source/__init__.py: re-exports the full public surface so every
  existing import path keeps working unchanged.

Dependency direction: protocol -> gossip -> live, strict and one-way.

Inside live.py, two structural cleanups:

- _accept_streams (243 lines, three nested concerns) decomposes into
  _negotiate_inbound_stream, _handle_gossipsub_inbound_stream,
  _setup_outbound_gossipsub_after_delay (was a nested closure), and
  _handle_reqresp_inbound_stream. The orchestrator becomes ~30 lines:
  accept loop plus protocol-id dispatch.
- _handle_inbound_quic_connection collapses into _handle_inbound_connection
  (the two were byte-identical except for one log line). _listen_quic
  now passes the unified handler.

Every existing docstring and inline comment is preserved verbatim from
main; only the structural pieces above changed. No external import path
moves; no test file is touched.

Verified:

- uvx tox -e all-checks (ruff, format, ty, codespell, mdformat) passes.
- tests/lean_spec/subspecs/networking/client and gossipsub tests
  (159 cases) pass.
