+++ src/lean_spec/subspecs/networking/client/event_source/__init__.py
@@ -1,0 +1,14 @@
+"""Network event source bridging transport to sync service."""
+
+from .gossip import GossipHandler, read_gossip_message
+from .live import LiveNetworkEventSource
+from .protocol import SUPPORTED_PROTOCOLS, EventSource, GossipMessageError
+
+__all__ = [
+    "SUPPORTED_PROTOCOLS",
+    "EventSource",
+    "GossipHandler",
+    "GossipMessageError",
+    "LiveNetworkEventSource",
+    "read_gossip_message",
+]
